convert user-timings to TS
