+++ static/src/javascripts/lib/user-timing.ts
@@ -1,0 +1,30 @@
+import performanceAPI from './window-performance';
+
+const timings: Record<string, number> = {};
+
+const getCurrentTime = (): number => performanceAPI.now();
+
+const markTime = (label: string): void => {
+	// eslint-disable-next-line @typescript-eslint/no-unnecessary-condition -- we’re checking out-of-spec browsers
+	if (performanceAPI && 'mark' in performanceAPI) {
+		performanceAPI.mark(label);
+	} else {
+		timings[label] = getCurrentTime();
+	}
+};
+
+// Returns the ms time when the mark was made.
+const getMarkTime = (label: string) => {
+	// eslint-disable-next-line @typescript-eslint/no-unnecessary-condition -- we’re checking out-of-spec browsers
+	if (performanceAPI && 'getEntriesByName' in performanceAPI) {
+		const perfMark = performanceAPI.getEntriesByName(label, 'mark');
+
+		if (perfMark[0] && 'startTime' in perfMark[0]) {
+			return perfMark[0].startTime;
+		}
+	} else if (label in timings) {
+		return timings[label];
+	}
+};
+
+export { markTime, getMarkTime, getCurrentTime };
